Session 121.2: Eliminate discovery env vars, add git-based repo-awareness
Replace PM_CONFIG_DIR and PM_SERVER_BIN environment variables with git-based
directory discovery. All processes now find .pm/ via `git rev-parse --show-toplevel`
with fallback chain (git → config.json → ~/.pm/).

Changes:
- Config::config_dir() uses 3-level fallback (git, config.json, global)
- Tauri directories moved from global macOS app_data_dir to repo-local .pm/
- user.json, server_dir, tauri_dir all now in <repo>/.pm/
- pm desktop command added to launch Tauri from CLI
- Child process cwd set to repo root (replaces PM_CONFIG_DIR env)
- Binary discovery refactored: sibling → .pm/bin/ → PATH
- Runtime config env vars (PM_SERVER_PORT, etc.) kept unchanged
- DRY violation fixed in port_file_info.rs
- All clippy warnings resolved

+++ backend/crates/pm-ws/tests/dependency_handler_tests.rs
@@ -481,7 +481,7 @@ async fn given_50_dependencies_when_create_51st_then_limit_error() {
             .bind(blocker_id.to_string())
             .bind(fixture.project_id.to_string())
             .bind(format!("Blocker {}", i))
-            .bind(i as i32 + 10)
+            .bind(i + 10)
             .bind(Utc::now().timestamp())
             .bind(Utc::now().timestamp())
             .bind(fixture.user_id.to_string())
